Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-scoping/host-functional-descendant-invalidation-expected.txt


FAIL CSS Test: element style is correctly updated for rule with :host(..) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
